+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
--
+- `nitrokey.trussed.admin_app.InitStatus`: add support for `EXT_FLASH_NEED_REFORMAT`
 
 [All Changes](https://github.com/Nitrokey/nitrokey-sdk-py/compare/v0.3.2...HEAD)