+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `nitrokey.trussed.admin_app.InitStatus`: add support for `EXT_FLASH_NEED_REFORMAT`
 - `nitrokey.nk3.updates`: return device status after an update
+- Use `poetry-core` v2 as build backend.
 
 [All Changes](https://github.com/Nitrokey/nitrokey-sdk-py/compare/v0.3.2...HEAD)